// src/interface/nota.interface.ts
// Generated by https://quicktype.io

export interface Nota {
  id:               string;
  matricula:        string;
  detalle: string;
  nota?: number;
  recomendacion?: string;
}

export interface NotaVO{
  id:               string;
  titulo: string;
  descripcion: string;
  matricula:        string;
  detalle: string;
  nota?: number;
  recomendacion?: string;
}